+++ tests/test_schema.py
@@ -66,6 +66,61 @@ def test_mlm_no_input_allowed_but_explicit_empty_array_required(
         pystac.validation.validate(mlm_item, validator=mlm_validator)
 
 
+@pytest.mark.parametrize(
+    "mlm_example",
+    ["item_basic.json"],
+    indirect=True,
+)
+def test_mlm_other_non_mlm_assets_allowed(
+    mlm_validator: STACValidator,
+    mlm_example: Dict[str, JSON],
+) -> None:
+    mlm_data = copy.deepcopy(mlm_example)
+    mlm_item = pystac.Item.from_dict(mlm_data)
+    pystac.validation.validate(mlm_item, validator=mlm_validator)  # self-check valid beforehand
+
+    mlm_data["assets"]["sample"] = {  # type: ignore
+        "type": "image/jpeg",
+        "href": "https://example.com/sample/output.jpg",
+        "roles": ["preview"],
+        "title": "Model Output Predictions Sample",
+    }
+    mlm_data["assets"]["model-cart"] = {  # type: ignore
+        "type": "text/markdown",
+        "href": "https://example.com/sample/model.md",
+        "roles": ["metadata"],
+        "title": "Model Cart",
+    }
+    mlm_item = pystac.Item.from_dict(mlm_data)
+    pystac.validation.validate(mlm_item, validator=mlm_validator)  # still valid
+
+
+@pytest.mark.parametrize(
+    "mlm_example",
+    ["item_basic.json"],
+    indirect=True,
+)
+def test_mlm_at_least_one_asset_model(
+    mlm_validator: STACValidator,
+    mlm_example: Dict[str, JSON],
+) -> None:
+    mlm_data = copy.deepcopy(mlm_example)
+    mlm_item = pystac.Item.from_dict(mlm_data)
+    pystac.validation.validate(mlm_item, validator=mlm_validator)  # self-check valid beforehand
+
+    mlm_data["assets"] = {  # needs at least 1 asset with role 'mlm:model'
+        "model": {
+            "type": "application/octet-stream; application=pytorch",
+            "href": "https://example.com/sample/checkpoint.pt",
+            "roles": ["checkpoint"],
+            "title": "Model Weights Checkpoint",
+        }
+    }
+    with pytest.raises(pystac.errors.STACValidationError):
+        mlm_item = pystac.Item.from_dict(mlm_data)
+        pystac.validation.validate(mlm_item, validator=mlm_validator)
+
+
 def test_model_metadata_to_dict(eurosat_resnet):
     assert eurosat_resnet.item.to_dict()